# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOLOS+ ESPORTZ clan management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all clan members
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, rejected, kicked]
        - name: role
          in: query
          schema:
            type: string
            enum: [OWNER, MANAGEMENT, TIER1, TIER2, TIER3, NEW_MEMBER]
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found

    patch:
      operationId: updateMember
      tags: [members]
      summary: Update member profile or stats (management only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete/kick member (management only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /members/{id}/approve:
    post:
      operationId: approveMember
      tags: [members]
      summary: Approve pending member application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/reject:
    post:
      operationId: rejectMember
      tags: [members]
      summary: Reject pending member application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/promote:
    post:
      operationId: promoteMember
      tags: [members]
      summary: Promote member to next tier or management role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleChange"
      responses:
        "200":
          description: Promoted member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/demote:
    post:
      operationId: demoteMember
      tags: [members]
      summary: Demote member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleChange"
      responses:
        "200":
          description: Demoted member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /members/profile:
    patch:
      operationId: updateMyProfile
      tags: [members]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, email, codmUsername, displayName, role, status, clanPoints, kdRatio, mvpCount, activityScore, scrimWins, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        codmUsername:
          type: string
        displayName:
          type: string
          description: "Clan-tagged username e.g. S²十NESS"
        whatsappNumber:
          type: ["string", "null"]
        customTag:
          type: ["string", "null"]
          description: "Custom role tag e.g. Admin, Clan Master, Sniper"
        role:
          type: string
          enum: [OWNER, MANAGEMENT, TIER1, TIER2, TIER3, NEW_MEMBER]
        status:
          type: string
          enum: [pending, active, rejected, kicked]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        clanPoints:
          type: integer
        kdRatio:
          type: number
        mvpCount:
          type: integer
        activityScore:
          type: integer
        scrimWins:
          type: integer
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        achievements:
          type: array
          items:
            type: string
        createdAt:
          type: string
    MemberUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [OWNER, MANAGEMENT, TIER1, TIER2, TIER3, NEW_MEMBER]
        status:
          type: string
          enum: [pending, active, rejected, kicked]
        clanPoints:
          type: integer
        kdRatio:
          type: number
        mvpCount:
          type: integer
        activityScore:
          type: integer
        scrimWins:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        bio:
          type: string
        avatarUrl:
          type: string
        whatsappNumber:
          type: string
        customTag:
          type: string
    RoleChange:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [OWNER, MANAGEMENT, TIER1, TIER2, TIER3, NEW_MEMBER]
...
    Notification:
      type: object
      required: [id, userId, type, title, message, read, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [approval, promotion, demotion, announcement, scrim, tournament, leaderboard, system]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
...