# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLOS+ ESPORTZ clan management platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateRole } from './memberUpdateRole';
import type { MemberUpdateStatus } from './memberUpdateStatus';

export interface MemberUpdate {
  role?: MemberUpdateRole;
  status?: MemberUpdateStatus;
  clanPoints?: number;
  kdRatio?: number;
  mvpCount?: number;
  activityScore?: number;
  scrimWins?: number;
}